requirements.txt
fastapi
python-multipart
PyMuPDF
Pillow
pytesseract
